Separated main into modules; fixed privacy access of rendering layer entities
