+++ examples/bouncing-square/index.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>Simple Square Bouncing Example</title>
+        <title>Bouncing Square Example - TypeScript render engine</title>
         <style>
             body,html {
                 padding: 0;
