+++ examples/index.html
@@ -6,10 +6,10 @@
         <p>Examples:</p>
         <ul>
             <li>
-                <a href="./pacman-animals/index.html">Pacman Example</a>
+                <a href="./bouncing-square/index.html">Square Bouncing Example</a>
             </li>
             <li>
-                <a href="./bouncing-square/index.html">Square Bouncing Example</a>
+                <a href="./pacman/index.html">Pacman Example</a>
             </li>
         </ul>
     </body>
