+++ examples/pacman/dog-and-cat-cover.jpg
