+++ examples/pacman/index.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>Pacman Animals Example</title>
+        <title>Pacman Example - TypeScript render engine</title>
         <style>
             body,html {
                 padding: 0;
