+++ examples/pacman/mod.d.ts
